Hard delete organization data (#1561)
We want to clear some unused orgs in IHCI prod (PATH and Screening data Wardha) to make the transition to Regions easier.

This adds a rake task to delete an org and its associated data.

An earlier attempt at cleanup for PATH didn't go correctly and we just deleted:

-  the PATH org,
- all facility groups belonging to PATH.

This went fine because we don't have a FK constraint between facilities and FGs. Currently there are ~130 facilities (and all the associated records, patients etc) with facility_group_id as nil which all belong to PATH. Hence the script has a small method to specifically cleanup PATH data.