Block-level sync fixes (#1876)
For the new apps that understand sync_region_id; they won't have the process token with sync_region_id available on the very time they sync (after the server gets deployed). This will cause the comparison for sync_region_modified? to fail and effectively trigger a forced resync. This should be avoided because there's no need to trigger this resync until we selectively enable block-level sync for users and/or all users eventually. When we do trigger block-sync for all users, the forced resync will be restricted to the patient data in their block; which should be smaller and less strenuous.

Makes an exception for a force-resync when the process token does not have a sync_region_id.