Optimize loading block_region_ids when syncing facilities
There are ~4k facilities in production, fetching the block_region itself
takes 2 queries + 1 query where we fetch the facility for the page.

This means a grand total of potentially 8000 region-related queries get
fire just for inserting a block_region_id.

This preloads the block_regions for all facilities that need to be
shipped and that brings us back to 4 queries for a pagesize of 1000.