Dont use let! ->
This makes all of our test state very hard to modify. For example if we want
to turn on regions_prep for these specs we need to be able to do that
via a flag in a test, but with the let! blocks its too late.

Also, using let! means we end up creating more objects then we need for
tests, which slows things down.