Show the entire tree down to accessible_facility_regions
We need to grab the lowest level a user has access to, and then walk our
way back up to properly show the hiearchy here..._and_ we need to know
if a user has access to parent regions or not. So using a very basic
permissions cache here to tell if the user has access to parent block /
district.