Regions report index listing (#1895)
* Spike out reusing the access tree presenter

* Dont use let! ->

This makes all of our test state very hard to modify. For example if we want
to turn on regions_prep for these specs we need to be able to do that
via a flag in a test, but with the let! blocks its too late.

Also, using let! means we end up creating more objects then we need for
tests, which slows things down.

* Grab org ids and FG ids to find admins via those models

Finding accessible_admins only via facilities feels too restrictive, and
won't find admins who happen to be created for a FG or an Org where
there aren't facilities created yet.

* Basic specs; accessible_blocks using path + region_type

* More direct way to grab the paths

* Simplify row collapse JS; allow non-async AdminAccess init

* Extract index tree partial

* Get index view working tree style

* Add ability to find accessible regions via source models

* Fix up old method names

* Show the entire tree down to accessible_facility_regions

We need to grab the lowest level a user has access to, and then walk our
way back up to properly show the hiearchy here..._and_ we need to know
if a user has access to parent regions or not. So using a very basic
permissions cache here to tell if the user has access to parent block /
district.

* Remove unnecessary regions_prep flag

* Fix spec

* Get readable SQL queries in our console logs again

* Handle auth better in regions controller

* linting

* Fix specs

* Add active_record_query_trace

see https://github.com/brunofacca/active-record-query-trace

* Check power user condition first

This will always be the fast condition to complete in the case of an
actual power user logged in, so we should check it first

* Preload the email_authentications on the current_admin

This prevents a TON of unnecessary email_auth loads...most of which hit
the AR cache, but they are still unnecessary and hurt performance and
also clog up the logs with all the unnecessary checks.

* linting