Region source slug mismatch (#1931)
**Story card:** [ch2182](https://app.clubhouse.io/simpledotorg/story/2182/region-source-slug-mismatch)

## Because

The links on the index page were broken when a region's slug didn't match its source's slug. The cause of breakage was: Link were crafted using the region's slug - but when clicked, the region lookup on the report page was using the URL param to search for _source_ records. This didn't matter in most cases, but is conceptually incorrect and can become a problem.

## This addresses

* Include the region lookup under the feature flag, so we directly look up regions by slug when feature flag is active
* Reuse the convenience method `accessible_block_regions` to perform block lookup and authorization