Auto deploy when on release tag and previous envs have passed
Requiring folks to go to the semaphore workflow page is an additional
manual step that doesn't feel necessary anymore.

This turns on auto_promote for release tags for prod envs if the
previous stages have passed.  The only other manual step here for
delivery would be the CHANGELOG piece, which we can automate next.

The deploying developer still needs to verify releases to the best of
their ability, of course.