Make sure we find correct model via the region slug
We need to make sure we find FacilityGroups and Facilities if region_reports is
disabled, and find Regions if region_reports is enabled. Otherwise slug
mismatches can result in records not being found, which ends up
returning authorization errors to users.

We also need to duck type source in some of the source models to make
sure the download links still work correctly.

This will all be resolved once all users are in the world of region
reports, and are list views _only_ show Regions and never FGs or
Facilities...so the links will have the correct slugs.