Add a Rachet to start decreasing our spec's reliance on let! (#2065)
* Add a Rachet to start decreasing our spec's reliance on let!

let! should almost never be used in specs, as it leads to lots of
implicit creation of test objects as well as making tests harder to read
and maintain.

This adds a rachet to help us in slowly decreasing the usage of let!. It
will fail if you add more let! usage, and also tells you to lower the
rachet count down if you have removed our usage of let!.

See https://thoughtbot.com/blog/lets-not and https://thoughtbot.com/blog/my-issues-with-let for more context on why let! is bad some alternatives to use.

* linting

* better regex

* Better error matches; bump up count