Exclude BPs from deleted patients from matview (#2133)
**Story card:** [ch2855](https://app.clubhouse.io/simpledotorg/story/2855/negative-reports-numbers)

## Because

If a patient is discarded but their BPs are not, their BPs are counted in reports numerators, but they are not counted in denominators. This is bad math, and leads to wonky results like negative numbers.

## This addresses

Updating the latest BPs per patient per month matview to exclude undiscarded BPs that belong to discarded patients.

The matview already excludes discarded BPs, and we rely on this behavior to circumvent some nasty ActiveRecord business. I think our options to protect against this were:
* ~~**Better data validations on BPs and patients:**~~ Since BP and patient data are synced asynchronously to us, I chose not to enforce any data validation rules on BPs and patients themselves
* ~~**Update the control and uncontrol queries to exclude discarded patient data**~~: I also chose not to do this because 1) the default scopes have trouble operating on our views, and 2) our assumption that only undiscarded data lives in the matview may be more widespread
* **Update the matview to ignore BPs that belong to discarded patients**: I think this is earliest layer where we can keep our data clean, so I've gone ahead with this.