Show drug stock report on success page (mobile app) (#2150)
* Begin spiking out drug stock report

* Wiring up with the actual report

* Show correct category; remove commented code

* Setup cache busting

* Generalize the patient css class for webview and dashboard

* Add the collection cache_key to drug stock reports cache_key

using the collection's cache_key method will ensure that the cache gets
bumped whne new drug stocks are entered...so we don't need to bypass the
cache manually anymore.  activerecord has the smarts builtin to handle
this for us

* Protocol should be https in dev too

This is primarily used for pointing the android app to your local
server, in which case you need to use HTTPS or Android won't even
connect.  Given our normal setup is via ngrok for the https tunnel, we
may as well default this to https here.

* linting and formatting

* Remove bust_cache params from specs

* remove dead code

* Rename to patient_days_css_class

as this isn't strictly background related anymore

* update label