Update BP control range to fall in line:
It will now return times that exactly represent the begin and end of the
BP control range to the exact second.  So queries can use the operands
you would expect for these sorts of queries, namely >= and <= ....the
previous inconsistency was confusing, where we would get the last date
of the day before the BP control range started.