Fix Patient#current_age (and flakey spec!)
We've had a flakey spec failure in the patient_exporter_spec that has
been bugging out for a long time now...see an exmample of the failure
here: https://simple.semaphoreci.com/jobs/393a5bee-61bb-4094-9097-756671673bd3

This was actually a real bug that just happens to pop up randomly based
on the current date and the date patients get from the factory.

Our prevoius current_age calculation did not take into account months
when returning the age based on date_of_birth, so we would get
intermittent off by one errors.  This fixes that, so the spec should be
stable now.