Do not include rxnorm code if it is not available (#2527)
**Story card:** [ch3545](https://app.clubhouse.io/simpledotorg/story/3545/missing-rxnorm-code-fix)

## Because

Some medications in our canonical list do not have an RxNorm code. When one of these medications is included in an import file, a validation error is thrown. 

## This addresses

Strip `rxnorm` from the params payload if it is missing. The param is optional in our API anyway. It just cannot be `nil`.