Deploy master to production if passing
This changes our semaphore config to auto promost the master branch to
production for all environments, assuming CI passes and previous
workflows have passed.

When we move to this sort of continous deployment, it doesn't really
make sense to cut tags for every release, because basically every merge
ref will become a release.

I've kept the logic around deploying tags that start with `release-` for
backwards compatiblity, for times when we want to cut releases by hand
still.