Get things matching more closely
Needed to account for the 'active range' for each region, and pass that
into the new query, otherwise we won't get the same results as the
previous query.  We need to know the range so that groupdate can fill in
0's for months without any counts, and we need to know the beginnign of
the range so that we don't get counts stretching back before the range
the user is asking for.