Simplify the error handling
We can set the context as tags for Sentry, so that it will roll things
up nicely in the Sentry UI.  Also, we don't need to capture the message
as an ivar, in fact the original exception is captured by Ruby for us as
the `cause` - and Sentry will show details on that.

I'm adding the original message into the custom message we have just for
a bit easier scanning when looking at errors in Sentry.