Initialize region service with a region (#2813)
**Story card:** [ch4413](https://app.clubhouse.io/simpledotorg/story/4413/my-facilities-bug)

Quick-fix for latest My Facilities bug. The problem surfaces when a facility and its region have different slugs (which is common). This makes everyone happy by initializing the `RegionService` with region records for now.

The follow-up robust fix would be to teach the `RegionService` to convert whatever it receives into region records as soon as possible, similar to how the `Repository` does today.