Bugfix: Encounter follow ups don't show up on same day (#2859)
**Story card:** [ch4414](https://app.clubhouse.io/simpledotorg/story/4414/et-zeros-in-progress-tab)

## Because

Daily follow up numbers on the progress tab number don't show up on the same day.

## This addresses

When a facility has both HTN and DM enabled, we fetch daily numbers by comparing `encounters.encountered_on`. However this is a date field, not a datetime field, and the date range comparison excludes today's results. This happens because groupdate tries to convert the time zone of the date. The [recommended way](https://github.com/ankane/groupdate#date-columns)  to avoid the time zone conversion is to pass in a  `time_zone: false` to groupdate.