Resolve logging difficulties
sigh, rails logging is a mess.

We use a combination of Ougai and Lograge to get structured log data
that sends nicely over to Datadog.  The problem is that they both muck
about with log formats, as does Rails TaggedLogging, and there can
easily be conflicts different libraries attempting to extend built in
formatters, or even Rails or Webpacker just doing surprising things with
log formats.

This consolidates our log patches into one file, which we require
_after_ rails but before Gems or app initialization.

I've tried to add comments to logging_extensions to try properly
document all the craziness here...and at least part of this can go away
with Rails 6.