Fix lk help screen (#2913)
**Story card:** [ch4977](https://app.shortcut.com/simpledotorg/story/4977)

## Because

LK was missing some support URL and number config

## This addresses

Ensuring the page doesn't break if we don't have support group config. It also cleans up some old code that used to support BD not having this config.

State of affairs:
* We had a dummy config for BD that just copied India
* We used CSS to hide the help section

We did this because the help screen was tough to work with while it was a giant locale string. Now that it's a proper ERB view, we can be smarter about this, hiding the section with conditionals if the necessary config is missing.