Update README.md (#2920)
* Update README.md

seeded admin account described on the documentation

* Update README.md

* update readme

Tightened up the wording on admin account created a bit.

* Update README.md (#2923)

* Add LK support info (#2919)

**Story card:** [ch4977](https://app.shortcut.com/simpledotorg/story/4977/sync-failure-in-sri-lanka-actionview-template-error-no-implicit-conversion-of-nil-into-string)

## Because

We got support info from Dr. Udara, so let's put it in the Sri Lanka help page.

* Fix overdue list download button (#2921)

* Add profiling RAILS_ENV and make it easy to run Datadog in local dev (#2914)

* Allow enabling Datadog via env var

for testing and tracing things in dev

* Manually enable rack mini profiler

because we don't need it in dev if we are using datadog and trying to do
more real world perf testing

* Better env var and comments

* remove old config

* Create a new profiling environment

For use in doing more 'production like' perf work and verification in
dev

* Build in v2 JSON manifest for dev / review environments

* Resolve logging difficulties

sigh, rails logging is a mess.

We use a combination of Ougai and Lograge to get structured log data
that sends nicely over to Datadog.  The problem is that they both muck
about with log formats, as does Rails TaggedLogging, and there can
easily be conflicts different libraries attempting to extend built in
formatters, or even Rails or Webpacker just doing surprising things with
log formats.

This consolidates our log patches into one file, which we require
_after_ rails but before Gems or app initialization.

I've tried to add comments to logging_extensions to try properly
document all the craziness here...and at least part of this can go away
with Rails 6.

* Create the formatter everytime, don't cache; better comments

* Simplify this bit of config

* remove extra stuff from profiling; clean up some noise

* Auto generate manifests for android_review (#2922)

* 4921 add diabetes treatment to medical history (#2906)

* 4921: add receiving_treatment_for_diabetes to medical history

Co-authored-by: Hari Mohanraj <hari.mohanraj89@gmail.com>
Co-authored-by: Prabhanshu Gupta <guptaprabhanshu1@gmail.com>
Co-authored-by: Rob Sanheim <rsanheim@gmail.com>
Co-authored-by: Kris Pethtel <kris.pethtel@gmail.com>

* username description updated

description for username and password updated,

Co-authored-by: Hari Mohanraj <hari.mohanraj89@gmail.com>
Co-authored-by: Prabhanshu Gupta <guptaprabhanshu1@gmail.com>
Co-authored-by: Rob Sanheim <rsanheim@gmail.com>
Co-authored-by: Kris Pethtel <kris.pethtel@gmail.com>