Consolidate refresh logic: make it possible to disable CONCURRENT
We need to be able to disable concurrent refreshes.  There are times
when we can't use concurrent refresh - for example, when the matviews
are completely empty after loading a fresh database.

Also, there may be times that doing a regular refresh will be more
performant and where the exclusive lock behavior doesn't matter (ie in
dev).