Fix two time zone issues; consolidate some setup
We need to execute reporting code in tests in the reporting time zone,
otherwise we get failures in certain times in certain time
zones...for example:

Tests are executing on a machine set to Central or Eastern time.  We are freezing a
date to Dec or Jan, so we are in CST or EST.  That means that starting
at 12:30 pm CST (so 12:30 pm developer time), IST time is one day ahead
(12:30 PM CST is 12:00 AM IST).

So tests that were written with the assumption of system time zone
matching the reporting timezone (which is true in India) will fail for
CI or for devs in US time zones.

By wrapping the code under test in the reporting_time_zone helper, we
get consistent behavior w/ time zones which also matches what happens in
production...because all our reports are generated inside controllers
that enforce IST timezones.