DataScript to update Bangladesh regions (#2918)
* Start data script generator

We gotta automate this process

* fix names and spec

* add logger

* Add initial BD script; update generators

* refactor

* Only delete community or small facilities

* Capture results

* Return results

* Initial CSV reading and processing

* Capture results

* Add CSV; add test CSV; more specs and simplify

* better spec

* clean up

* Refactor and get specs passing

* USC is also a community facility

* Shouldn't have added this as a regular migration

* data migration, not regular

* Fix intermittent spec - freeze the entire example

* linting

* Need to delete facilities with no size as well

* Destroy Region when a Facility is destroyed

* Better result hash; make sure we find the right Region

* linting

* docs

* Make sure we clean up orphaned FacilityGroups

* Incorporate some PR feedback

* Attach DHIS2 identifiers

* more accurate description

* Remove new users

* linting

* Extract to DataScript

* Enforce read only for dry_run

* clean up

* refactor and more specs

* remove focus

* Add failsafe for non BD

* more assertions

* lint

* Turn off read only mode after all done

* check for failed deletes

* better error handling

* Test w/ real CSV to help find region bug

* Fix issue with duplicate regions

Need to find regions by name by searching amongst direct children,
otherwise we get issues with returning incorrect regions in where names
are the same.

* fix dry run

* go back

* Consolidate refresh logic: make it possible to disable CONCURRENT

We need to be able to disable concurrent refreshes.  There are times
when we can't use concurrent refresh - for example, when the matviews
are completely empty after loading a fresh database.

Also, there may be times that doing a regular refresh will be more
performant and where the exclusive lock behavior doesn't matter (ie in
dev).

* scaffold for running thru this

* Consolidate refresh logic: make it possible to disable CONCURRENT

We need to be able to disable concurrent refreshes.  There are times
when we can't use concurrent refresh - for example, when the matviews
are completely empty after loading a fresh database.

Also, there may be times that doing a regular refresh will be more
performant and where the exclusive lock behavior doesn't matter (ie in
dev).

* Shouldn't have added a module here

* Don't need these transactions here

* Dont need outer tx here:

The tx wrapping is done at the refresh method level now

* Fix two time zone issues; consolidate some setup

We need to execute reporting code in tests in the reporting time zone,
otherwise we get failures in certain times in certain time
zones...for example:

Tests are executing on a machine set to Central or Eastern time.  We are freezing a
date to Dec or Jan, so we are in CST or EST.  That means that starting
at 12:30 pm CST (so 12:30 pm developer time), IST time is one day ahead
(12:30 PM CST is 12:00 AM IST).

So tests that were written with the assumption of system time zone
matching the reporting timezone (which is true in India) will fail for
CI or for devs in US time zones.

By wrapping the code under test in the reporting_time_zone helper, we
get consistent behavior w/ time zones which also matches what happens in
production...because all our reports are generated inside controllers
that enforce IST timezones.

* Simplify and also comment on our time zone handling here

* disable concurrent refresh for this load

* needs to be conditional

* c'mon

* More harness to get this working

* Fix this include

* clean up

* clean up

* lint

* Remove standardrb exceptions

we no longer need this, as everything is in order

* Fix spec: ensure constant lookup works

* Fix spec to properly capture abort

* Destroy related regions only when they have no children

* Smarter cascading region delete; clean up empty blocks

* tweak output

* tweak script

* Bit more logging

* Bit more logging

* didn't need this