Remove facility performance feature
This was never enabled in any prod environments, and I believe it was
put on the back-burner because of many other higher priority epics
getting prioritized. The code has fallen out of date w/ all the recent
reporting pipeline changes, so we can remove it to avoid unnecessary
maintenance.

We can always add it back from this commit and
bring it up to date w/ Repository code if we get it re-prioritized and
into development.