Add Sandbox to the Demo manifest (#3332)
**Story card:** [sc-6861](https://app.shortcut.com/simpledotorg/story/6861)

## Because

We want to allow google play reviewers to test the full production app in "demo mode".

## This addresses

Adding a "Sandbox" country to the demo manifest. This is a trial run for adding a "Sandbox" country to the production manifest, thereby allowing Google Play reviewers to select it and do all the review they need in a test environment.