Demo country in production (#3341)
**Story card:** [sc-6861](https://app.shortcut.com/simpledotorg/story/6861/fix-google-play-policy-issue)

## Because

We need to make the production app browsable in a "demo" mode for Google Play app reviewers to give us the 👍 Without this approval, we cannot make any releases to the production Simple app.

## This addresses

Add a "Demo" country to the production manifest that points at ~Sandbox~ Demo. Users who select this country (eg. Google Play reviewers) can self sign-up and connect to Sandbox. Upon re-sign-in, they can use the fixed OTP of 000000, since Sandbox has fixed OTPs.

This approach avoids the following, which is good:

* Introducing conditional OTP logic into the application
* Adding test data to a production server
* Introducing any "demo mode" into the app that can be switched to, resulting in confusion for real users.

This is a follow-up PR to #3340 and #3332