Wire up follow_ups v2 in progress tab (#3316)
* Wire feature flag for follow ups v2

* Fix feature flag name

* Better spec - showing that we do need distinct in the matview

When a follow up is created from different users, the underlying matview
reporting_patient_follow_ups will track each distinctly. So we need to
change our roll up to be smarter about this case and count only distinct
patients per month.

* Need to count distinct patients for correct follow up counts

* We need both feature flags for this path right now

* Add some sum methods for registrations and follow ups

* Make sure we get the unique index on reporting_facility_state_groups

* Make it easier to grab all results necessary in the controller

this is still convoluted, but starting to get to something better

* more explicit todo

* Simplify this API

* Split out the diagnosis_code

* Linting

* Spike out a better sum approach

* Add specs for the all time totals

* comments

* More refactoring

* Use the new counts / total_counts api

* Use a null object for cases where we have no data

* remove debugging

* remove dead code

* WIP

* Remove v2 component; render things inline to simplify

also add some debugging to the view so we can see what is what

* lint

* Introduce FacilityProgressDimension:

a value type to capture the particular dimenion of data we want to grab
from the pipeline and show in the progress table.  Along the lines of:

* indicator (registrations or diabetes right now)
* diagnosis (can be all)
* gender (can be all)

* Build up dimensions and let the component render one each

Much simpler and nicer to define the dimensions first and then let the
component render one each.

* Need monthly_counts here too

* Get specs passing; clean up dead code

* Remove dead code

* Lint

* remove safe operation navigator

don't think we should need it

* linter

* Display zeros and with delimiter

* Remove debugging

* better error messages