Run the exporter every day for now (#3592)
**Story card:** [sc-5903](https://app.shortcut.com/simpledotorg/story/5903/turn-on-maharashtra-dhis2-export-flag-in-india-for-mahashtra)

This PR adds the MH DHIS2 export to our daily cron tasks. We choose to keep this as a daily task so we can quickly iterate with minimal code changes during the initial setup. Once this works, we can move this to a monthly cadence - maybe when we hook up to a production server.

For now, the data is being sent to a test server managed by the Maharashtra HISP. The following other dependencies have already been taken care of:
* [x] DHIS2 IDs added to the 10 Maharashtra facilities involved in the integration
* [x] The feature flag `maharashtra_dhis2_export` has been enabled in India production
* [x] The necessary code that exports gender-disaggregated data elements to Maharashtra has been merged