Task to refresh region names (#3606)
**Story card:** [sc-7612](https://app.shortcut.com/simpledotorg/story/7612/refresh-region-names-of-seed-data-in-india-demo)

## Because

We want to make the names of places look Indian in India demo. We had updated the "fake_names" csv, but need to effect that change.

## This PR

Introduces a one-off script to execute this rename in Sandbox first (to do a prod-like test) and then in India demo.

Of course, if we ever purged and reseeded any other demo environment, the names would become Indian, but we can deal with that later.