Remove the regions refresh script (#3615)
**Story card:** [sc-7612](https://app.shortcut.com/simpledotorg/story/7612/refresh-region-names-of-seed-data-in-india-demo)

We introduced a one-off script to refresh sandbox and India demo region names. The script has been executed successfully, so we can remove it now. The script was also not perfect, as it led to [some integrity failures](https://simpledotorg.slack.com/archives/C013WT1QY5V/p1647410422867399), so no reason to keep it around.

## This addresses

Removes the `lib/tasks/refresh_region_names.rb` script that was introduced in #3606