Drug Stock API  Enhancement (#3698)
**Story card:** [sc-8103](https://app.shortcut.com/simpledotorg/story/8103/update-drugstockreminderapi-to-include-drugstockformurl)

The mobile client needs to know the URL where the drug stock submission form is for its facility/user. On the progress tab, we bake the URL into the progress tab when rendering serverside, but for the home screen reminder, the app doesn't have that URL easily available.

This PR adds the submission URL to the drug stocks API endpoint that the client will hit before showing the reminder.

It also allows the "back" button to be hidden. We want to do this when the drug stock form is opened from the home screen.