Consolidate hypertension reports (#3726)
**Story card:** [sc-8281](https://app.shortcut.com/simpledotorg/story/8281/consolidate-hypertension-reports)

## Because

Diabetes reports are going to be launched in a new "Diabetes" tab on the reports page. This makes the information architecture of the reports a little weird, with several HTN tabs and one DM tab.

## This addresses

This PR consolidates the Overview, Details, and Cohort Reports pages into a single "Hypertension" page in the reports section. Here are some additional notes about the change.

* The restructuring is only effective when the `diabetes_management_reports` feature flag is turned on. Several controller actions and views and stuff can be deleted when we are removing the feature flag.
* We have chosen to simply display both monthly and quarterly cohort reports one after the other, instead of using a dropdown to toggle between them
* There is some substantial overlap of controller code between the different sections of the hypertension report. We will schedule a cleanup for later. We want to make the restructure right away so we can get ahead of the launch of the DM reports.