Added a disabled DOWNLOADS button to DM reports.  (#3760)
No story. It's just confusing that we have a Downloads dropdown with HTN downloads on the DM Reports. See screenshot below for the disabled button.

I also removed a completely unnecessary print-only menu that is now extra confusing with the DM Reports. I removed everything after "Apr-2022 Report" in this code (it's totally not needed).