CSS prep
using split-bars and :root to access the same classes inside and outside the 'split-bars' class. Otherwise I have to create double the number of classes.

I think it is better to use a single class to define a single colour across the board to prevent missing a key and a bar colour.