Quick hack to fix scrolling on wide and long screens
On windows horizontal scroll bar is at the bottom of the list.
Where the list is long users need to scroll to the bottom of the list so they can access the horizontal scroll bar.
This places the scroll bars at a browser view.
Not a pretty fix - but functional