Fix translation locale selection (#5153)
**Story card:** Bug-fix

This PR does 3 things:
1. Prunes `config.i18n.available_locales` to only contain keys which are
in the translation backends.
2. Move from `compatible_language_from` to
`language_region_compatible_from` method of http_accept_language gem,
which is more correct.
1. Simple server was ignoring region part of translation in favour of
first occurring primary language, which is incorrect. (For ex, `bn-IN`
got inferred as `bn-BD` because that was defined first in config)
3. Change all underscores (en_BD) to hyphens (en-BD), because hyphen is
the convention.

## Previous PR cause a production issue because:

1. `config.i18n.available_locales` contained a lot of locales that were
not present in translation files as backend.
3. `en-IN` was defaulting to `en-IN` by
`language_region_compatible_from` method, because
`config.i18n.available_locales` contained `en_IN` key, and a backend for
this wasn't present.
1. This was causing an error: `I18n::InvalidLocale: "en-IN" is not a
valid locale`
4. This wasn't an issue before this PR, because previously we were using
this method: `compatible_language_from`, which inferred `en-IN` as `en`.