facility bulk upload: accept business identifiers
We add a new field to the bulk upload template that if populated, would
also create a business identifier mapping to the facility.

This is an important first step that will let us accept external
organisations that want to export their data and see the reports on the
Simple dashboard.

We also correct a mistake in the template which says that the facility
size field is optional. This is untrue since 2cdf24c and this was likely
something that was missed when that change was made.