Fetch dogstatsd host from env if available (#5189)
**Story card:**
[10793](https://app.shortcut.com/simpledotorg/story/10793)

## Because

Dogstatsd metrics are missing in k8s environments

## This addresses

Updates the setup configuration of the DataDog agent for Kubernetes
environments. The key change is the setup of the dogstatsd host, which
should now be set to the DataDog agent service name, unlike in VM
deployments where it is accessible over localhost.