Allow import of patients via the bulk import API (#5229)
**Story card:**
https://app.shortcut.com/simpledotorg/story/10962/add-ingestion-for-patients
[sc-10962]

## Because

We are currently mocking bulk API imports. We now actually accept real
resources if they are patient resources.

## This addresses

* Adds an import job and import type dispatcher.
* Adds another feature flag to have everything be a mock. This flag will
be removed when the implementation of all resources is complete. We
should enable this flag in sandbox, while partner organisations are
still in the process of sending their facility data.
* Moves the `ImportUser` class out of the `PatientImport` namespace,
since we are going to use this user for our FHIR imports.
* Updates some swagger schema errors.

## Test instructions

Follow the instructions in this document to insert a resource in
Sandbox:

https://docs.google.com/document/d/1f-H474ozfNBjiISMd7IOLQtZREVJh86u6JBnw4235qw/edit?usp=sharing