bulk facility upload: fix incorrect duplicate check (#5243)
The bulk facility upload tool always reported that the business
identifiers were duplicated. This is because of a one-letter typo in our
code. We fix this and add a test case to avoid this in the future.

Explanation: we are plucking `identifiers` from
`FacilityBusinessIdentifiers`, but the attribute name is a singular
`identifier`. This incorrect pluck creates a map of `nil`s and causes
the uniqueness check to always fail.