fix incorrect drug stock instruction source
She/he who presently runs an rswag generator will get dunked by a stray
diff that is not desired. It is an old instruction in the specimen
example for the API docs that should not have seen the light of day,
much like the pav bhaji dosa. We banish this bothersome stray back the
eternal void.